+++ src/main/java/backend/wal/wal/nextwal/domain/aggregate/Category.java
@@ -1,6 +1,6 @@
-package backend.wal.onboard.nextwal.domain.aggregate;
+package backend.wal.wal.nextwal.domain.aggregate;
 
-import backend.wal.onboard.common.WalCategoryType;
+import backend.wal.wal.common.domain.WalCategoryType;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
