+++ src/main/java/backend/wal/wal/nextwal/domain/repository/ItemRepository.java
@@ -1,7 +1,7 @@
-package backend.wal.onboard.nextwal.domain.repository;
+package backend.wal.wal.nextwal.domain.repository;
 
-import backend.wal.onboard.common.WalCategoryType;
-import backend.wal.onboard.nextwal.domain.aggregate.Item;
+import backend.wal.wal.common.domain.WalCategoryType;
+import backend.wal.wal.nextwal.domain.aggregate.Item;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ItemRepository extends JpaRepository<Item, Long> {
